:truck: Intento de reorganizacion con diseño atomico
