+++ app/src/atoms/toaster.tsx
@@ -5,7 +5,7 @@ import {
   ToastProvider,
   ToastTitle,
   ToastViewport,
-} from "@/components/ui/toast"
+} from "@/atoms/toast"
 import { useToast } from "@/hooks/use-toast"
 
 export function Toaster() {
