+++ app/src/main.tsx
@@ -1,4 +1,4 @@
-import { Toaster } from "@/components/ui/toaster"
+import { Toaster } from "@/atoms/toaster.tsx"
 import React from "react"
 import ReactDOM from "react-dom/client"
 import App from "./App.tsx"
