+++ app/src/molecules/EmpiricalSpectrum.tsx
