+++ app/src/molecules/InferenceBoxGraph.tsx
