+++ app/src/molecules/ReferenceLampRange.tsx
