+++ app/src/molecules/ReferenceLampSpectrum.tsx
