+++ app/src/organisms/BBImageEditor.tsx
@@ -3,7 +3,7 @@ import type { ReactNode } from "react"
 import { Spectrum } from "@/enums/Spectrum"
 import { useEffect, useRef, useState } from "react"
 import { Pane, ResizablePanes } from "resizable-panes-react"
-import { Button } from "./ui/button"
+import { Button } from "../atoms/button"
 
 const spectrumColors: Record<Spectrum, string> = {
     [Spectrum.Lamp]: "red",
