+++ app/src/organisms/InferenceForm.tsx
@@ -1,9 +1,9 @@
 import { useGlobalStore } from "@/hooks/use-global-store"
 import { CustomError, legendreAlgoritm, linearRegression, piecewiseLinearRegression } from "@/lib/utils"
 import { useEffect, useMemo, useState } from "react"
-import { ErrorScatterGraph } from "./ErrorScatterGraph"
-import { InferenceBoxGraph } from "./InferenceBoxGraph"
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "./ui/card"
+import { ErrorScatterGraph } from "../molecules/ErrorScatterGraph"
+import { InferenceBoxGraph } from "../molecules/InferenceBoxGraph"
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "../atoms/card"
 
 interface InferenceOption {
     id: number
