+++ app/src/organisms/NavigationProgressBar.tsx
@@ -1,8 +1,8 @@
 import { useGlobalStore } from "@/hooks/use-global-store"
 import { ChevronLeftIcon, ChevronRightIcon } from "@heroicons/react/24/outline"
 import { useState } from "react"
-import { Button } from "./ui/button"
-import "./css/ProgressBar.css"
+import { Button } from "../atoms/button"
+import "@/css/ProgressBar.css"
 
 interface ProgressBarProps {
     value: number
