+++ app/src/pages/StepSpectrumSegmentation.tsx
@@ -2,9 +2,9 @@ import type { BoundingBox } from "@/interfaces/BoundingBox"
 import type { StepProps } from "@/interfaces/StepProps"
 import type { ChangeEvent } from "react"
 import { useState } from "react"
-import { BBImageEditor } from "./BBImageEditor"
-import { Button } from "./ui/button"
-import { Uploader } from "./Uploader"
+import { Button } from "../atoms/button"
+import { Uploader } from "../molecules/Uploader"
+import { BBImageEditor } from "../organisms/BBImageEditor"
 
 type LoadingState = "waiting" | "processing" | "finished" | "error"
